style: extracted to a separate class
